R library
Dear all I uploaded a library of R functions (and called it
SpeciesGeoCodeR) to the R folder. It can do 4 things:
1. use objects passed on from python and produce outputtables and
graphics
2. In addition one can slo perform the point in polygon test entirely in
R, if people prefer R
3. be use to produce fancy maps of point and polygon datasets
4. subset the WWF shape file to polygons for realms, biomes or
ecoregions and convert them into a list of points.

The readme files contains a rough description for all functions, the
readme_mats file some information on the input needed from python.
There are some examples of possible input and outputfiles in the
examples/examples_! folder. Looking forward to talking to you about this
on thursday.